+++ Assets/Ecosim/Scenes/Gameplay/_Scripts/Entity/Component/Storage/CellSnapshot.cs
@@ -5,9 +5,9 @@ namespace Ecosim
     [Serializable]
     public class CellSnapshot
     {
-        public readonly ushort Amount;
+        public readonly int Amount;
 
-        public CellSnapshot(ushort amount)
+        public CellSnapshot(int amount)
         {
             Amount = amount;
         }
